implement warp::reply::file, warp::header::conditionals, and move warp::fs::Conditionals to warp::header
