+++ src/filters/header.rs
@@ -8,12 +8,14 @@ use std::convert::Infallible;
 use std::str::FromStr;
 
 use futures_util::future;
-use headers::{Header, HeaderMapExt};
+use headers::{Header, HeaderMapExt, IfModifiedSince, IfRange, IfUnmodifiedSince, LastModified, Range};
 use http::header::HeaderValue;
-use http::HeaderMap;
+use http::{HeaderMap, StatusCode};
+use hyper::Body;
 
 use crate::filter::{filter_fn, filter_fn_one, Filter, One};
 use crate::reject::{self, Rejection};
+use crate::reply::Response;
 
 /// Create a `Filter` that tries to parse the specified header.
 ///
@@ -227,4 +229,96 @@ pub fn value(
 /// ```
 pub fn headers_cloned() -> impl Filter<Extract = One<HeaderMap>, Error = Infallible> + Copy {
     filter_fn_one(|route| future::ok(route.headers().clone()))
+}
+
+/// Create a `Filter` that returns a container for conditional headers
+///
+/// # Example
+/// ```
+/// use warp::Filter;
+/// use warp::fs::Conditionals;
+///
+/// let headers = warp::header::conditionals()
+///     .and_then(|conditionals: Conditionals| {
+///         warp::reply::file("index.html", conditionals)
+///     });
+/// ```
+pub fn conditionals() -> impl Filter<Extract=One<Conditionals>, Error = Infallible> + Copy {
+    optional2()
+        .and(optional2())
+        .and(optional2())
+        .and(optional2())
+        .map(
+            |if_modified_since, if_unmodified_since, if_range, range| Conditionals {
+                if_modified_since,
+                if_unmodified_since,
+                if_range,
+                range,
+            },
+        )
+}
+
+
+/// Utilized conditional headers to determine response-content
+#[derive(Debug, Default)]
+pub struct Conditionals {
+    if_modified_since: Option<IfModifiedSince>,
+    if_unmodified_since: Option<IfUnmodifiedSince>,
+    if_range: Option<IfRange>,
+    range: Option<Range>,
+}
+
+pub(crate) enum ConditionalBody {
+    NoBody(Response),
+    WithBody(Option<Range>),
+}
+
+impl Conditionals {
+    pub(crate) fn check(self, last_modified: Option<LastModified>) -> ConditionalBody {
+        if let Some(since) = self.if_unmodified_since {
+            let precondition = last_modified
+                .map(|time| since.precondition_passes(time.into()))
+                .unwrap_or(false);
+
+            tracing::trace!(
+                "if-unmodified-since? {:?} vs {:?} = {}",
+                since,
+                last_modified,
+                precondition
+            );
+            if !precondition {
+                let mut res = Response::new(Body::empty());
+                *res.status_mut() = StatusCode::PRECONDITION_FAILED;
+                return ConditionalBody::NoBody(res);
+            }
+        }
+
+        if let Some(since) = self.if_modified_since {
+            tracing::trace!(
+                "if-modified-since? header = {:?}, file = {:?}",
+                since,
+                last_modified
+            );
+            let unmodified = last_modified
+                .map(|time| !since.is_modified(time.into()))
+                // no last_modified means its always modified
+                .unwrap_or(false);
+            if unmodified {
+                let mut res = Response::new(Body::empty());
+                *res.status_mut() = StatusCode::NOT_MODIFIED;
+                return ConditionalBody::NoBody(res);
+            }
+        }
+
+        if let Some(if_range) = self.if_range {
+            tracing::trace!("if-range? {:?} vs {:?}", if_range, last_modified);
+            let can_range = !if_range.is_modified(None, last_modified.as_ref());
+
+            if !can_range {
+                return ConditionalBody::WithBody(None);
+            }
+        }
+
+        ConditionalBody::WithBody(self.range)
+    }
 }
